syntax/ext: collect the ast building traits into a single trait.
