+++ src/libsyntax/ext/pipes/proto.rs
@@ -11,7 +11,8 @@
 use ast;
 use codemap::span;
 use ext::base::ExtCtxt;
-use ext::pipes::ast_builder::{append_types, ext_ctxt_ast_builder, path};
+use ext::build::AstBuilder;
+use ext::pipes::ast_builder::{append_types, path};
 
 #[deriving(Eq)]
 pub enum direction { send, recv }
@@ -93,7 +94,7 @@ pub impl state_ {
 
     /// Returns the type that is used for the messages.
     fn to_ty(&self, cx: @ExtCtxt) -> @ast::Ty {
-        cx.ty_path_ast_builder
+        cx.ty_path
             (path(~[cx.ident_of(self.name)],self.span).add_tys(
                 cx.ty_vars(&self.generics.ty_params)))
     }
